RaspGateway/requirements.txt
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
netifaces==0.10.6
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
simple-websocket==1.0.0
socketio==0.2.1
urllib3==2.2.1
websocket-client==1.7.0
wsproto==1.2.0
